Moving traits and Rank generic class
